[ADD] simple yet lovely permutator

+++ srcs/infect/disasm.c
@@ -194,7 +194,7 @@ static void	disasm_instruction(const void *code, size_t codelen, uint32_t *src, 
 	*dst |= i.dst & 0xff000000;
 }
 
-uint64_t	disasm(const void *code, size_t codelen, struct s_instruction *buf, size_t buflen)
+size_t	disasm(const void *code, size_t codelen, struct s_instruction *buf, size_t buflen)
 {
 	void			*p_code = (void*)code;
 	struct s_instruction	*p_buf  = (void*)buf;
